Introduce `MutableAccessAwareProject`
