+++ subprojects/core/src/integTest/groovy/org/gradle/api/invocation/GradleLifecycleAllprojectsIntegrationTest.groovy
@@ -17,96 +17,18 @@
 package org.gradle.api.invocation
 
 import org.gradle.integtests.fixtures.AbstractIntegrationSpec
-import org.gradle.test.precondition.Requires
-import org.gradle.test.preconditions.IntegTestPreconditions
 
-@Requires(IntegTestPreconditions.NotIsolatedProjects)
 class GradleLifecycleAllprojectsIntegrationTest extends AbstractIntegrationSpec {
 
-    def 'lifecycle.allprojects is not executed again after project evaluation'() {
+    def 'lifecycle.allprojects is executed before lifecycle.beforeProject'() {
         given:
         settingsFile << """
             rootProject.name = 'root'
-            gradle.lifecycle.allprojects { project ->
-                println "lifecycle.allprojects for \${project.name}"
+            gradle.lifecycle.beforeProject {
+                println "lifecycle.beforeProject for \${name}"
             }
-            include(":a")
-        """
-
-        file("a/build.gradle") << "println 'a'"
-
-        buildFile << """
-            println 'root'
-            allprojects { project ->
-                println "allprojects for \${project.name}"
-            }
-            afterEvaluate {
-                allprojects { project ->
-                    println "afterEvaluate allprojects for \${project.name}"
-                }
-            }
-        """
-
-        when:
-        run "help", "-q"
-
-        then:
-        outputContains """
-lifecycle.allprojects for root
-root
-allprojects for root
-lifecycle.allprojects for a
-allprojects for a
-afterEvaluate allprojects for root
-afterEvaluate allprojects for a
-a
-"""
-    }
-
-    def 'lifecycle.allprojects is executed only once for a project'() {
-        given:
-        settingsFile << """
-            rootProject.name = 'root'
-            gradle.lifecycle.allprojects { project ->
-                println "lifecycle.allprojects for \${project.name}"
-            }
-            include(":a")
-        """
-
-        file("a/build.gradle") << "println 'a'"
-
-        buildFile << """
-            allprojects { project ->
-                println "allprojects for \${project.name}"
-            }
-            subprojects { project ->
-                println "subprojects for \${project.name}"
-            }
-        """
-
-        when:
-        run "help", "-q"
-
-        then:
-        outputContains """
-lifecycle.allprojects for root
-allprojects for root
-lifecycle.allprojects for a
-allprojects for a
-subprojects for a
-a
-"""
-    }
-
-    def 'lifecycle.allprojects is executed before gradle.lifecycle.beforeProject'() {
-        given:
-        settingsFile << """
-            rootProject.name = 'root'
-            gradle.lifecycle.beforeProject { project ->
-                println "lifecycle.beforeProject for \${project.name}"
-            }
-            gradle.lifecycle.allprojects { project ->
-                println "lifecycle.allprojects for \${project.name}"
+            gradle.lifecycle.allprojects {
+                println "lifecycle.allprojects for \${name}"
             }
             include(":a")
         """
@@ -122,172 +44,5 @@ lifecycle.beforeProject for root
 lifecycle.allprojects for a
 lifecycle.beforeProject for a
 """
-    }
-
-    def 'lifecycle.allprojects is executed before project.#api'() {
-        given:
-        settingsFile << """
-            rootProject.name = 'root'
-
-            gradle.lifecycle.allprojects { project ->
-                project.ext.foo = "bar"
-            }
-            include(":a")
-            include(":b")
-        """
-
-        file("a/build.gradle") << ""
-        file("b/build.gradle") << ""
-
-        buildFile << """
-            $invocation { project ->
-                println "foo = \${project.foo} for \${project.name}"
-            }
-        """
-
-        when:
-        run "help", "-q"
-
-        then:
-        outputContains expectedOutput
-
-        where:
-        api                | invocation                            | expectedOutput
-        "allprojects"      | "allprojects"                         | allprojectsExpectedOutput
-        "subprojects"      | "subprojects"                         | subprojectsExpectedOutput
-        "project"          | "project(':a')"                       | projectExpectedOutput
-        "getAllprojects"   | "getAllprojects().forEach"            | allprojectsExpectedOutput
-        "getSubprojects"   | "getSubprojects().forEach"            | subprojectsExpectedOutput
-        "getChildProjects" | "getChildProjects().values().forEach" | subprojectsExpectedOutput
-    }
-
-    private static def allprojectsExpectedOutput = """
-foo = bar for root
-foo = bar for a
-foo = bar for b
-"""
-    private static def subprojectsExpectedOutput = """
-foo = bar for a
-foo = bar for b
-"""
-    private static def projectExpectedOutput = """
-foo = bar for a
-"""
-
-    def 'lifecycle.allprojects is executed eagerly only if a mutable state of a project touched by using project.#api'() {
-        given:
-        settingsFile << """
-            rootProject.name = 'root'
-            gradle.lifecycle.allprojects { project ->
-                println "lifecycle.allprojects for \${project.name}"
-                project.ext.foo = "bar"
-            }
-            include(":a")
-            include(":b")
-        """
-
-        file("a/build.gradle") << "println 'a'"
-        file("b/build.gradle") << """
-            println 'b'
-            println "foo = \${foo} for b"
-        """
-
-        buildFile << """
-            println 'root'
-            $invocation { project ->
-                if (project.getName() == 'a') {
-                    ${mutableStateAccessFor("a")}
-                } else {
-                    println "Immutable state access for \${project.getName()}"
-                }
-            }
-        """
-
-        when:
-        run "help", "-q"
-
-        then:
-        outputContains expectedOutput
-
-        where:
-        api                | invocation                            | expectedOutput
-        "allprojects"      | "allprojects"                         | mutableStateAccessProjectAllprojectsExpectedOutput
-        "subprojects"      | "subprojects"                         | mutableStateAccessProjectSubprojectsExpectedOutput
-        "getAllprojects"   | "getAllprojects().forEach"            | mutableStateAccessProjectAllprojectsExpectedOutput
-        "getSubprojects"   | "getSubprojects().forEach"            | mutableStateAccessProjectSubprojectsExpectedOutput
-        "getChildProjects" | "getChildProjects().values().forEach" | mutableStateAccessProjectSubprojectsExpectedOutput
-    }
-
-    private static def mutableStateAccessProjectAllprojectsExpectedOutput = """
-lifecycle.allprojects for root
-root
-Immutable state access for root
-Before mutable state access for a
-lifecycle.allprojects for a
-After mutable state access for a(foo = bar)
-Immutable state access for b
-a
-lifecycle.allprojects for b
-b
-foo = bar for b
-"""
-    private static def mutableStateAccessProjectSubprojectsExpectedOutput = """
-lifecycle.allprojects for root
-root
-Before mutable state access for a
-lifecycle.allprojects for a
-After mutable state access for a(foo = bar)
-Immutable state access for b
-a
-lifecycle.allprojects for b
-b
-foo = bar for b
-"""
-
-    def 'lifecycle.allprojects is executed eagerly only if a mutable state of a project touched by using gradle.allprojects'() {
-        given:
-        settingsFile << """
-            rootProject.name = 'root'
-            gradle.lifecycle.allprojects { project ->
-                println "lifecycle.allprojects for \${project.getName()}"
-                project.ext.foo = "bar"
-            }
-            gradle.allprojects { project ->
-                if (project.getName() == 'a') {
-                    ${mutableStateAccessFor("a")}
-                } else {
-                    println "Immutable state access for \${project.getName()}"
-                }
-            }
-            include(":a")
-        """
-
-        file("a/build.gradle") << "println 'a'"
-
-        buildFile << """
-            println 'root'
-        """
-
-        when:
-        run "help", "-q"
-
-        then:
-        outputContains """
-Immutable state access for root
-Before mutable state access for a
-lifecycle.allprojects for a
-After mutable state access for a(foo = bar)
-lifecycle.allprojects for root
-root
-a
-"""
-    }
-
-    private static String mutableStateAccessFor(String project) {
-        return """
-            println "Before mutable state access for $project"
-            def value = project.foo
-            println "After mutable state access for $project(foo = \$value)"
-        """
     }
 }
