+++ subprojects/core/src/main/java/org/gradle/api/internal/project/MutableStateAccessAwareProject.java
@@ -1,0 +1,1166 @@
+/*
+ * Copyright 2024 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.api.internal.project;
+
+import groovy.lang.Closure;
+import groovy.lang.GroovyObjectSupport;
+import groovy.lang.MissingPropertyException;
+import groovy.lang.Script;
+import org.gradle.api.Action;
+import org.gradle.api.AntBuilder;
+import org.gradle.api.InvalidUserDataException;
+import org.gradle.api.NamedDomainObjectContainer;
+import org.gradle.api.NamedDomainObjectFactory;
+import org.gradle.api.PathValidation;
+import org.gradle.api.Project;
+import org.gradle.api.ProjectEvaluationListener;
+import org.gradle.api.Task;
+import org.gradle.api.UnknownProjectException;
+import org.gradle.api.artifacts.dsl.ArtifactHandler;
+import org.gradle.api.artifacts.dsl.DependencyFactory;
+import org.gradle.api.artifacts.dsl.DependencyHandler;
+import org.gradle.api.artifacts.dsl.DependencyLockingHandler;
+import org.gradle.api.artifacts.dsl.RepositoryHandler;
+import org.gradle.api.component.SoftwareComponentContainer;
+import org.gradle.api.file.ConfigurableFileCollection;
+import org.gradle.api.file.ConfigurableFileTree;
+import org.gradle.api.file.CopySpec;
+import org.gradle.api.file.DeleteSpec;
+import org.gradle.api.file.FileTree;
+import org.gradle.api.file.ProjectLayout;
+import org.gradle.api.file.SyncSpec;
+import org.gradle.api.internal.GradleInternal;
+import org.gradle.api.internal.ProcessOperations;
+import org.gradle.api.internal.artifacts.configurations.RoleBasedConfigurationContainerInternal;
+import org.gradle.api.internal.file.FileOperations;
+import org.gradle.api.internal.file.FileResolver;
+import org.gradle.api.internal.initialization.ClassLoaderScope;
+import org.gradle.api.internal.initialization.ScriptHandlerInternal;
+import org.gradle.api.internal.plugins.ExtensionContainerInternal;
+import org.gradle.api.internal.plugins.PluginManagerInternal;
+import org.gradle.api.internal.tasks.TaskContainerInternal;
+import org.gradle.api.internal.tasks.TaskDependencyFactory;
+import org.gradle.api.logging.Logger;
+import org.gradle.api.logging.LoggingManager;
+import org.gradle.api.model.ObjectFactory;
+import org.gradle.api.plugins.ObjectConfigurationAction;
+import org.gradle.api.plugins.PluginContainer;
+import org.gradle.api.project.IsolatedProject;
+import org.gradle.api.provider.Property;
+import org.gradle.api.provider.Provider;
+import org.gradle.api.provider.ProviderFactory;
+import org.gradle.api.resources.ResourceHandler;
+import org.gradle.api.tasks.WorkResult;
+import org.gradle.configuration.ConfigurationTargetIdentifier;
+import org.gradle.configuration.project.ProjectConfigurationActionContainer;
+import org.gradle.groovy.scripts.ScriptSource;
+import org.gradle.internal.Cast;
+import org.gradle.internal.logging.StandardOutputCapture;
+import org.gradle.internal.metaobject.BeanDynamicObject;
+import org.gradle.internal.metaobject.DynamicInvokeResult;
+import org.gradle.internal.metaobject.DynamicObject;
+import org.gradle.internal.metaobject.DynamicObjectUtil;
+import org.gradle.internal.model.ModelContainer;
+import org.gradle.internal.model.RuleBasedPluginListener;
+import org.gradle.internal.service.ServiceRegistry;
+import org.gradle.internal.service.scopes.ServiceRegistryFactory;
+import org.gradle.model.internal.registry.ModelRegistry;
+import org.gradle.normalization.InputNormalizationHandler;
+import org.gradle.normalization.internal.InputNormalizationHandlerInternal;
+import org.gradle.process.ExecResult;
+import org.gradle.process.ExecSpec;
+import org.gradle.process.JavaExecSpec;
+import org.gradle.util.Path;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import java.io.File;
+import java.net.URI;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.concurrent.Callable;
+
+public abstract class MutableStateAccessAwareProject extends GroovyObjectSupport implements ProjectInternal {
+
+    protected final ProjectInternal delegate;
+
+    protected MutableStateAccessAwareProject(ProjectInternal delegate) {
+        this.delegate = delegate;
+    }
+
+    abstract void onMutableStateAccess();
+
+    @Override
+    public abstract boolean equals(Object obj);
+
+    @Override
+    public int hashCode() {
+        return delegate.hashCode();
+    }
+
+    @Nullable
+    protected Object invokeMethodOnThisBean(String methodName, Object[] args) {
+        DynamicObject thisBean = new BeanDynamicObject(this).withNotImplementsMissing();
+        DynamicInvokeResult result = thisBean.tryInvokeMethod(methodName, args);
+        return result.isFound() ? result.getValue() : null;
+    }
+
+    @Nullable
+    protected Object getPropertyOnThisBean(String propertyName) {
+        DynamicObject thisBean = new BeanDynamicObject(this).withNotImplementsMissing();
+        DynamicInvokeResult result = thisBean.tryGetProperty(propertyName);
+        return result.isFound() ? result.getValue() : null;
+    }
+
+    @Override
+    @Nullable
+    public Object invokeMethod(String name, Object args) {
+        Object[] varargs = Cast.uncheckedNonnullCast(args);
+        Object thisBeanResult = invokeMethodOnThisBean(name, varargs);
+        if (thisBeanResult != null) {
+            return thisBeanResult;
+        }
+
+        onMutableStateAccess();
+
+        DynamicObject delegateDynamicObject = DynamicObjectUtil.asDynamicObject(delegate);
+        DynamicInvokeResult delegateResult = delegateDynamicObject.tryInvokeMethod(name, varargs);
+
+        if (delegateResult.isFound()) {
+            return delegateResult.getValue();
+        }
+
+        throw delegateDynamicObject.methodMissingException(name, varargs);
+    }
+
+    @Override
+    @Nullable
+    public Object getProperty(String propertyName) {
+        Object thisBeanResult = getPropertyOnThisBean(propertyName);
+        if (thisBeanResult != null) {
+            return thisBeanResult;
+        }
+
+        onMutableStateAccess();
+
+        DynamicObject delegateDynamicObject = DynamicObjectUtil.asDynamicObject(delegate);
+        DynamicInvokeResult delegateResult = delegateDynamicObject.tryGetProperty(propertyName);
+
+        if (delegateResult.isFound()) {
+            return delegateResult.getValue();
+        }
+
+        throw delegateDynamicObject.getMissingProperty(propertyName);
+    }
+
+    @Nullable
+    @Override
+    public ProjectInternal getParent() {
+        return delegate.getParent();
+    }
+
+    @Override
+    public String getName() {
+        return delegate.getName();
+    }
+
+    @Override
+    public String getDisplayName() {
+        return delegate.getDisplayName();
+    }
+
+    @Nullable
+    @Override
+    public String getDescription() {
+        onMutableStateAccess();
+        return delegate.getDescription();
+    }
+
+    @Override
+    public void setDescription(@Nullable String description) {
+        onMutableStateAccess();
+        delegate.setDescription(description);
+    }
+
+    @Override
+    public Object getGroup() {
+        onMutableStateAccess();
+        return delegate.getGroup();
+    }
+
+    @Override
+    public void setGroup(Object group) {
+        onMutableStateAccess();
+        delegate.setGroup(group);
+    }
+
+    @Override
+    public Object getVersion() {
+        onMutableStateAccess();
+        return delegate.getVersion();
+    }
+
+    @Override
+    public void setVersion(Object version) {
+        onMutableStateAccess();
+        delegate.setVersion(version);
+    }
+
+    @Override
+    public Object getStatus() {
+        onMutableStateAccess();
+        return delegate.getStatus();
+    }
+
+    @Override
+    public void setStatus(Object status) {
+        onMutableStateAccess();
+        delegate.setStatus(status);
+    }
+
+    @Nullable
+    @Override
+    public ProjectInternal getParent(ProjectInternal referrer) {
+        return delegate.getParent(referrer);
+    }
+
+    @Override
+    public ProjectInternal getRootProject() {
+        return delegate.getRootProject();
+    }
+
+    @Override
+    public File getRootDir() {
+        return delegate.getRootDir();
+    }
+
+
+    @Override
+    @SuppressWarnings("deprecation")
+    public File getBuildDir() {
+        onMutableStateAccess();
+        return delegate.getBuildDir();
+    }
+
+    @SuppressWarnings("deprecation")
+    @Override
+    public void setBuildDir(File path) {
+        onMutableStateAccess();
+        delegate.setBuildDir(path);
+    }
+
+    @SuppressWarnings("deprecation")
+    @Override
+    public void setBuildDir(Object path) {
+        onMutableStateAccess();
+        delegate.setBuildDir(path);
+    }
+
+    @Override
+    public File getBuildFile() {
+        return delegate.getBuildFile();
+    }
+
+    @Override
+    public ProjectInternal getRootProject(ProjectInternal referrer) {
+        return delegate.getRootProject(referrer);
+    }
+
+    @Override
+    public Project evaluate() {
+        return delegate.evaluate();
+    }
+
+    @Override
+    public ProjectInternal bindAllModelRules() {
+        return delegate.bindAllModelRules();
+    }
+
+    @Override
+    public TaskContainerInternal getTasks() {
+        onMutableStateAccess();
+        return delegate.getTasks();
+    }
+
+    @Override
+    public void subprojects(Action<? super Project> action) {
+        delegate.subprojects(action);
+    }
+
+    @Override
+    public void subprojects(Closure configureClosure) {
+        delegate.subprojects(configureClosure);
+    }
+
+    @Override
+    public void allprojects(Action<? super Project> action) {
+        delegate.allprojects(action);
+    }
+
+    @Override
+    public void allprojects(Closure configureClosure) {
+        delegate.allprojects(configureClosure);
+    }
+
+    @Override
+    public void beforeEvaluate(Action<? super Project> action) {
+        delegate.beforeEvaluate(action);
+    }
+
+    @Override
+    public void afterEvaluate(Action<? super Project> action) {
+        delegate.afterEvaluate(action);
+    }
+
+    @Override
+    public void beforeEvaluate(Closure closure) {
+        delegate.beforeEvaluate(closure);
+    }
+
+    @Override
+    public void afterEvaluate(Closure closure) {
+        delegate.afterEvaluate(closure);
+    }
+
+    @Override
+    public boolean hasProperty(String propertyName) {
+        onMutableStateAccess();
+        return delegate.hasProperty(propertyName);
+    }
+
+    @Override
+    public Map<String, ?> getProperties() {
+        onMutableStateAccess();
+        return delegate.getProperties();
+    }
+
+    @Nullable
+    @Override
+    public Object property(String propertyName) throws MissingPropertyException {
+        onMutableStateAccess();
+        return delegate.property(propertyName);
+    }
+
+    @Nullable
+    @Override
+    public Object findProperty(String propertyName) {
+        onMutableStateAccess();
+        return delegate.findProperty(propertyName);
+    }
+
+    @Override
+    public Logger getLogger() {
+        onMutableStateAccess();
+        return delegate.getLogger();
+    }
+
+    @Override
+    public ScriptSource getBuildScriptSource() {
+        return delegate.getBuildScriptSource();
+    }
+
+    @Override
+    public ProjectInternal project(String path) throws UnknownProjectException {
+        return delegate.project(path);
+    }
+
+    @Override
+    public Project project(String path, Closure configureClosure) {
+        return delegate.project(path, configureClosure);
+    }
+
+    @Override
+    public Project project(String path, Action<? super Project> configureAction) {
+        return delegate.project(path, configureAction);
+    }
+
+    @Override
+    public Map<Project, Set<Task>> getAllTasks(boolean recursive) {
+        onMutableStateAccess();
+        return delegate.getAllTasks(recursive);
+    }
+
+    @Override
+    public Set<Task> getTasksByName(String name, boolean recursive) {
+        onMutableStateAccess();
+        return delegate.getTasksByName(name, recursive);
+    }
+
+    @Override
+    public File getProjectDir() {
+        return delegate.getProjectDir();
+    }
+
+    @Override
+    public File file(Object path) {
+        onMutableStateAccess();
+        return delegate.file(path);
+    }
+
+    @Override
+    public File file(Object path, PathValidation validation) throws InvalidUserDataException {
+        onMutableStateAccess();
+        return delegate.file(path, validation);
+    }
+
+    @Override
+    public URI uri(Object path) {
+        onMutableStateAccess();
+        return delegate.uri(path);
+    }
+
+    @Override
+    public String relativePath(Object path) {
+        onMutableStateAccess();
+        return delegate.relativePath(path);
+    }
+
+    @Override
+    public ConfigurableFileCollection files(Object... paths) {
+        onMutableStateAccess();
+        return delegate.files(paths);
+    }
+
+    @Override
+    public ConfigurableFileCollection files(Object paths, Closure configureClosure) {
+        onMutableStateAccess();
+        return delegate.files(paths, configureClosure);
+    }
+
+    @Override
+    public ConfigurableFileCollection files(Object paths, Action<? super ConfigurableFileCollection> configureAction) {
+        onMutableStateAccess();
+        return delegate.files(paths, configureAction);
+    }
+
+    @Override
+    public ConfigurableFileTree fileTree(Object baseDir) {
+        onMutableStateAccess();
+        return delegate.fileTree(baseDir);
+    }
+
+    @Override
+    public ConfigurableFileTree fileTree(Object baseDir, Closure configureClosure) {
+        onMutableStateAccess();
+        return delegate.fileTree(baseDir, configureClosure);
+    }
+
+    @Override
+    public ConfigurableFileTree fileTree(Object baseDir, Action<? super ConfigurableFileTree> configureAction) {
+        onMutableStateAccess();
+        return delegate.fileTree(baseDir, configureAction);
+    }
+
+    @Override
+    public ConfigurableFileTree fileTree(Map<String, ?> args) {
+        onMutableStateAccess();
+        return delegate.fileTree(args);
+    }
+
+    @Override
+    public FileTree zipTree(Object zipPath) {
+        onMutableStateAccess();
+        return delegate.zipTree(zipPath);
+    }
+
+    @Override
+    public FileTree tarTree(Object tarPath) {
+        onMutableStateAccess();
+        return delegate.tarTree(tarPath);
+    }
+
+    @Override
+    public <T> Provider<T> provider(Callable<? extends T> value) {
+        onMutableStateAccess();
+        return delegate.provider(value);
+    }
+
+    @Override
+    public ProviderFactory getProviders() {
+        onMutableStateAccess();
+        return delegate.getProviders();
+    }
+
+    @Override
+    public ObjectFactory getObjects() {
+        onMutableStateAccess();
+        return delegate.getObjects();
+    }
+
+    @Override
+    public ProjectLayout getLayout() {
+        onMutableStateAccess();
+        return delegate.getLayout();
+    }
+
+    @Override
+    public File mkdir(Object path) {
+        onMutableStateAccess();
+        return delegate.mkdir(path);
+    }
+
+    @Override
+    public boolean delete(Object... paths) {
+        onMutableStateAccess();
+        return delegate.delete(paths);
+    }
+
+    @Override
+    public WorkResult delete(Action<? super DeleteSpec> action) {
+        onMutableStateAccess();
+        return delegate.delete(action);
+    }
+
+    @Override
+    public ExecResult javaexec(Closure closure) {
+        onMutableStateAccess();
+        return delegate.javaexec(closure);
+    }
+
+    @Override
+    public ExecResult javaexec(Action<? super JavaExecSpec> action) {
+        onMutableStateAccess();
+        return delegate.javaexec(action);
+    }
+
+    @Override
+    public ExecResult exec(Closure closure) {
+        onMutableStateAccess();
+        return delegate.exec(closure);
+    }
+
+    @Override
+    public ExecResult exec(Action<? super ExecSpec> action) {
+        onMutableStateAccess();
+        return delegate.exec(action);
+    }
+
+    @Override
+    public String absoluteProjectPath(String path) {
+        return delegate.absoluteProjectPath(path);
+    }
+
+    @Override
+    public String relativeProjectPath(String path) {
+        return delegate.relativeProjectPath(path);
+    }
+
+    @Override
+    public AntBuilder getAnt() {
+        onMutableStateAccess();
+        return delegate.getAnt();
+    }
+
+    @Override
+    public AntBuilder createAntBuilder() {
+        onMutableStateAccess();
+        return delegate.createAntBuilder();
+    }
+
+    @Override
+    public AntBuilder ant(Closure configureClosure) {
+        onMutableStateAccess();
+        return delegate.ant(configureClosure);
+    }
+
+    @Override
+    public AntBuilder ant(Action<? super AntBuilder> configureAction) {
+        onMutableStateAccess();
+        return delegate.ant(configureAction);
+    }
+
+    @Override
+    public ProjectInternal project(ProjectInternal referrer, String path) throws UnknownProjectException {
+        return delegate.project(referrer, path);
+    }
+
+    @Override
+    public ProjectInternal project(ProjectInternal referrer, String path, Action<? super Project> configureAction) {
+        return delegate.project(referrer, path, configureAction);
+    }
+
+    @Nullable
+    @Override
+    public ProjectInternal findProject(String path) {
+        return delegate.findProject(path);
+    }
+
+    @Nullable
+    @Override
+    public ProjectInternal findProject(ProjectInternal referrer, String path) {
+        return delegate.findProject(referrer, path);
+    }
+
+    @Override
+    public Set<? extends ProjectInternal> getSubprojects(ProjectInternal referrer) {
+        return delegate.getSubprojects(referrer);
+    }
+
+    @Override
+    public void subprojects(ProjectInternal referrer, Action<? super Project> configureAction) {
+        delegate.subprojects(referrer, configureAction);
+    }
+
+    @Override
+    public Map<String, Project> getChildProjects() {
+        return delegate.getChildProjects();
+    }
+
+    @Override
+    public void setProperty(String name, @Nullable Object value) throws MissingPropertyException {
+        onMutableStateAccess();
+        delegate.setProperty(name, value);
+    }
+
+    @Override
+    public IsolatedProject getIsolated() {
+        return delegate.getIsolated();
+    }
+
+    @Override
+    public Set<Project> getAllprojects() {
+        return delegate.getAllprojects();
+    }
+
+    @Override
+    public Set<Project> getSubprojects() {
+        return delegate.getSubprojects();
+    }
+
+    @Override
+    public Task task(String name) throws InvalidUserDataException {
+        onMutableStateAccess();
+        return delegate.task(name);
+    }
+
+    @Override
+    public Task task(Map<String, ?> args, String name) throws InvalidUserDataException {
+        onMutableStateAccess();
+        return delegate.task(args, name);
+    }
+
+    @Override
+    public Task task(Map<String, ?> args, String name, Closure configureClosure) {
+        onMutableStateAccess();
+        return delegate.task(args, name, configureClosure);
+    }
+
+    @Override
+    public Task task(String name, Closure configureClosure) {
+        onMutableStateAccess();
+        return delegate.task(name, configureClosure);
+    }
+
+    @Override
+    public Task task(String name, Action<? super Task> configureAction) {
+        onMutableStateAccess();
+        return delegate.task(name, configureAction);
+    }
+
+    @Override
+    public String getPath() {
+        return delegate.getPath();
+    }
+
+    @Nullable
+    @Override
+    public ProjectIdentifier getParentIdentifier() {
+        return delegate.getParentIdentifier();
+    }
+
+    @Override
+    public String getBuildTreePath() {
+        return delegate.getBuildTreePath();
+    }
+
+    @Override
+    public List<String> getDefaultTasks() {
+        onMutableStateAccess();
+        return delegate.getDefaultTasks();
+    }
+
+    @Override
+    public void setDefaultTasks(List<String> defaultTasks) {
+        onMutableStateAccess();
+        delegate.setDefaultTasks(defaultTasks);
+    }
+
+    @Override
+    public void defaultTasks(String... defaultTasks) {
+        onMutableStateAccess();
+        delegate.defaultTasks(defaultTasks);
+    }
+
+    @Override
+    public Project evaluationDependsOn(String path) throws UnknownProjectException {
+        onMutableStateAccess();
+        return delegate.evaluationDependsOn(path);
+    }
+
+    @Override
+    public void evaluationDependsOnChildren() {
+        onMutableStateAccess();
+        delegate.evaluationDependsOnChildren();
+    }
+
+    @Override
+    public Map<String, Project> getChildProjectsUnchecked() {
+        return delegate.getChildProjectsUnchecked();
+    }
+
+    @Override
+    public Set<? extends ProjectInternal> getAllprojects(ProjectInternal referrer) {
+        return delegate.getAllprojects(referrer);
+    }
+
+    @Override
+    public void allprojects(ProjectInternal referrer, Action<? super Project> configureAction) {
+        delegate.allprojects(referrer, configureAction);
+    }
+
+    @Override
+    public DynamicObject getInheritedScope() {
+        return delegate.getInheritedScope();
+    }
+
+    @Override
+    public GradleInternal getGradle() {
+        onMutableStateAccess();
+        return delegate.getGradle();
+    }
+
+    @Override
+    public LoggingManager getLogging() {
+        onMutableStateAccess();
+        return delegate.getLogging();
+    }
+
+    @Override
+    public Object configure(Object object, Closure configureClosure) {
+        onMutableStateAccess();
+        return delegate.configure(object, configureClosure);
+    }
+
+    @Override
+    public Iterable<?> configure(Iterable<?> objects, Closure configureClosure) {
+        onMutableStateAccess();
+        return delegate.configure(objects, configureClosure);
+    }
+
+    @Override
+    public <T> Iterable<T> configure(Iterable<T> objects, Action<? super T> configureAction) {
+        onMutableStateAccess();
+        return delegate.configure(objects, configureAction);
+    }
+
+    @Override
+    public RepositoryHandler getRepositories() {
+        onMutableStateAccess();
+        return delegate.getRepositories();
+    }
+
+    @Override
+    public void repositories(Closure configureClosure) {
+        onMutableStateAccess();
+        delegate.repositories(configureClosure);
+    }
+
+    @Override
+    public DependencyHandler getDependencies() {
+        onMutableStateAccess();
+        return delegate.getDependencies();
+    }
+
+    @Override
+    public void dependencies(Closure configureClosure) {
+        onMutableStateAccess();
+        delegate.dependencies(configureClosure);
+    }
+
+    @Override
+    public DependencyFactory getDependencyFactory() {
+        onMutableStateAccess();
+        return delegate.getDependencyFactory();
+    }
+
+    @Override
+    public ProjectEvaluationListener getProjectEvaluationBroadcaster() {
+        return delegate.getProjectEvaluationBroadcaster();
+    }
+
+    @Override
+    public void addRuleBasedPluginListener(RuleBasedPluginListener listener) {
+        delegate.addRuleBasedPluginListener(listener);
+    }
+
+    @Override
+    public void prepareForRuleBasedPlugins() {
+        delegate.prepareForRuleBasedPlugins();
+    }
+
+    @Override
+    public FileResolver getFileResolver() {
+        return delegate.getFileResolver();
+    }
+
+    @Override
+    public TaskDependencyFactory getTaskDependencyFactory() {
+        return delegate.getTaskDependencyFactory();
+    }
+
+    @Override
+    public ServiceRegistry getServices() {
+        return delegate.getServices();
+    }
+
+    @Override
+    public ServiceRegistryFactory getServiceRegistryFactory() {
+        return delegate.getServiceRegistryFactory();
+    }
+
+    @Override
+    public StandardOutputCapture getStandardOutputCapture() {
+        return delegate.getStandardOutputCapture();
+    }
+
+    @Override
+    public ProjectStateInternal getState() {
+        onMutableStateAccess();
+        return delegate.getState();
+    }
+
+    @Override
+    public <T> NamedDomainObjectContainer<T> container(Class<T> type) {
+        onMutableStateAccess();
+        return delegate.container(type);
+    }
+
+    @Override
+    public <T> NamedDomainObjectContainer<T> container(Class<T> type, NamedDomainObjectFactory<T> factory) {
+        onMutableStateAccess();
+        return delegate.container(type, factory);
+    }
+
+    @Override
+    public <T> NamedDomainObjectContainer<T> container(Class<T> type, Closure factoryClosure) {
+        onMutableStateAccess();
+        return delegate.container(type, factoryClosure);
+    }
+
+    @Override
+    public ExtensionContainerInternal getExtensions() {
+        onMutableStateAccess();
+        return delegate.getExtensions();
+    }
+
+    @Override
+    public ResourceHandler getResources() {
+        onMutableStateAccess();
+        return delegate.getResources();
+    }
+
+    @Override
+    public SoftwareComponentContainer getComponents() {
+        onMutableStateAccess();
+        return delegate.getComponents();
+    }
+
+    @Override
+    public void components(Action<? super SoftwareComponentContainer> configuration) {
+        onMutableStateAccess();
+        delegate.components(configuration);
+    }
+
+    @Override
+    public ProjectConfigurationActionContainer getConfigurationActions() {
+        return delegate.getConfigurationActions();
+    }
+
+    @Override
+    public ModelRegistry getModelRegistry() {
+        return delegate.getModelRegistry();
+    }
+
+    @Override
+    public ClassLoaderScope getClassLoaderScope() {
+        return delegate.getClassLoaderScope();
+    }
+
+    @Override
+    public ClassLoaderScope getBaseClassLoaderScope() {
+        return delegate.getBaseClassLoaderScope();
+    }
+
+    @Override
+    public void setScript(Script script) {
+        delegate.setScript(script);
+    }
+
+    @Override
+    public void addDeferredConfiguration(Runnable configuration) {
+        delegate.addDeferredConfiguration(configuration);
+    }
+
+    @Override
+    public void fireDeferredConfiguration() {
+        delegate.fireDeferredConfiguration();
+    }
+
+    @Override
+    public Path identityPath(String name) {
+        return delegate.identityPath(name);
+    }
+
+    @Override
+    public Path projectPath(String name) {
+        return delegate.projectPath(name);
+    }
+
+    @Nonnull
+    @Override
+    public Path getProjectPath() {
+        return delegate.getProjectPath();
+    }
+
+    @Nullable
+    @Override
+    public ProjectInternal getProject() {
+        return this;
+    }
+
+    @Override
+    public ModelContainer<?> getModel() {
+        return delegate.getModel();
+    }
+
+    @Override
+    public Path getBuildPath() {
+        return delegate.getBuildPath();
+    }
+
+    @Override
+    public boolean isScript() {
+        return delegate.isScript();
+    }
+
+    @Override
+    public boolean isRootScript() {
+        return delegate.isRootScript();
+    }
+
+    @Override
+    public boolean isPluginContext() {
+        return delegate.isPluginContext();
+    }
+
+    @Override
+    public Path getIdentityPath() {
+        return delegate.getIdentityPath();
+    }
+
+    @Nullable
+    @Override
+    public ProjectEvaluationListener stepEvaluationListener(ProjectEvaluationListener listener, Action<ProjectEvaluationListener> action) {
+        return delegate.stepEvaluationListener(listener, action);
+    }
+
+    @Override
+    public ProjectState getOwner() {
+        return delegate.getOwner();
+    }
+
+    @Override
+    public InputNormalizationHandlerInternal getNormalization() {
+        onMutableStateAccess();
+        return delegate.getNormalization();
+    }
+
+    @Override
+    public void normalization(Action<? super InputNormalizationHandler> configuration) {
+        onMutableStateAccess();
+        delegate.normalization(configuration);
+    }
+
+    @Override
+    public void dependencyLocking(Action<? super DependencyLockingHandler> configuration) {
+        onMutableStateAccess();
+        delegate.dependencyLocking(configuration);
+    }
+
+    @Override
+    public DependencyLockingHandler getDependencyLocking() {
+        onMutableStateAccess();
+        return delegate.getDependencyLocking();
+    }
+
+    @Override
+    public ScriptHandlerInternal getBuildscript() {
+        onMutableStateAccess();
+        return delegate.getBuildscript();
+    }
+
+    @Override
+    public void buildscript(Closure configureClosure) {
+        onMutableStateAccess();
+        delegate.buildscript(configureClosure);
+    }
+
+    @Override
+    public WorkResult copy(Closure closure) {
+        onMutableStateAccess();
+        return delegate.copy(closure);
+    }
+
+    @Override
+    public WorkResult copy(Action<? super CopySpec> action) {
+        onMutableStateAccess();
+        return delegate.copy(action);
+    }
+
+    @Override
+    public CopySpec copySpec(Closure closure) {
+        onMutableStateAccess();
+        return delegate.copySpec(closure);
+    }
+
+    @Override
+    public CopySpec copySpec(Action<? super CopySpec> action) {
+        onMutableStateAccess();
+        return delegate.copySpec(action);
+    }
+
+    @Override
+    public CopySpec copySpec() {
+        onMutableStateAccess();
+        return delegate.copySpec();
+    }
+
+    @Override
+    public WorkResult sync(Action<? super SyncSpec> action) {
+        onMutableStateAccess();
+        return delegate.sync(action);
+    }
+
+    @Override
+    public DetachedResolver newDetachedResolver() {
+        return delegate.newDetachedResolver();
+    }
+
+    @Override
+    public Property<Object> getInternalStatus() {
+        onMutableStateAccess();
+        return delegate.getInternalStatus();
+    }
+
+    @Override
+    public RoleBasedConfigurationContainerInternal getConfigurations() {
+        onMutableStateAccess();
+        return delegate.getConfigurations();
+    }
+
+    @Override
+    public void configurations(Closure configureClosure) {
+        onMutableStateAccess();
+        delegate.configurations(configureClosure);
+    }
+
+    @Override
+    public ArtifactHandler getArtifacts() {
+        onMutableStateAccess();
+        return delegate.getArtifacts();
+    }
+
+    @Override
+    public void artifacts(Closure configureClosure) {
+        onMutableStateAccess();
+        delegate.artifacts(configureClosure);
+    }
+
+    @Override
+    public void artifacts(Action<? super ArtifactHandler> configureAction) {
+        onMutableStateAccess();
+        delegate.artifacts(configureAction);
+    }
+
+    @SuppressWarnings("deprecation")
+    @Override
+    public org.gradle.api.plugins.Convention getConvention() {
+        onMutableStateAccess();
+        return delegate.getConvention();
+    }
+
+    @Override
+    public int depthCompare(Project otherProject) {
+        return delegate.depthCompare(otherProject);
+    }
+
+    @Override
+    public int getDepth() {
+        return delegate.getDepth();
+    }
+
+    @Override
+    public int compareTo(@Nonnull Project o) {
+        return delegate.compareTo(o);
+    }
+
+    @Override
+    public FileOperations getFileOperations() {
+        return delegate.getFileOperations();
+    }
+
+    @Override
+    public ProcessOperations getProcessOperations() {
+        return delegate.getProcessOperations();
+    }
+
+    @Override
+    public PluginContainer getPlugins() {
+        onMutableStateAccess();
+        return delegate.getPlugins();
+    }
+
+    @Override
+    public void apply(Closure closure) {
+        onMutableStateAccess();
+        delegate.apply(closure);
+    }
+
+    @Override
+    public void apply(Action<? super ObjectConfigurationAction> action) {
+        onMutableStateAccess();
+        delegate.apply(action);
+    }
+
+    @Override
+    public void apply(Map<String, ?> options) {
+        onMutableStateAccess();
+        delegate.apply(options);
+    }
+
+    @Override
+    public PluginManagerInternal getPluginManager() {
+        onMutableStateAccess();
+        return delegate.getPluginManager();
+    }
+
+    @Override
+    public ConfigurationTargetIdentifier getConfigurationTargetIdentifier() {
+        return delegate.getConfigurationTargetIdentifier();
+    }
+}
